fix(all): 修复场景部分 bug
