+++ frontend/src/views/api-test/management/components/management/index.vue
@@ -56,7 +56,6 @@
 </template>
 
 <script setup lang="ts">
-  import { onBeforeRouteLeave } from 'vue-router';
   import { cloneDeep } from 'lodash-es';
 
   import MsEditableTab from '@/components/pure/ms-editable-tab/index.vue';
@@ -70,7 +69,6 @@
   import { getProtocolList } from '@/api/modules/api-test/common';
   import { useI18n } from '@/hooks/useI18n';
   import useLeaveTabUnSaveCheck from '@/hooks/useLeaveTabUnSaveCheck';
-  import useModal from '@/hooks/useModal';
   import useAppStore from '@/store/modules/app';
   import { hasAnyPermission } from '@/utils/permission';
 
@@ -98,7 +96,6 @@
   }>();
   const appStore = useAppStore();
   const { t } = useI18n();
-  const { openModal } = useModal();
 
   const setActiveApi: ((params: RequestParam) => void) | undefined = inject('setActiveApi');
