+++ frontend/src/views/api-test/scenario/components/step/createAction/stepInsertStepTrigger.vue
@@ -7,7 +7,12 @@
       position="br"
       @popup-visible-change="handleActionTriggerChange"
     >
-      <MsButton :id="step.uniqueId" type="icon" class="ms-tree-node-extra__btn !mr-[4px]" @click="emit('click')">
+      <MsButton
+        :id="`trigger-${step.uniqueId}`"
+        type="icon"
+        class="ms-tree-node-extra__btn !mr-[4px]"
+        @click="emit('click')"
+      >
         <MsIcon type="icon-icon_add_outlined" size="14" class="text-[var(--color-text-4)]" />
       </MsButton>
       <template #content>
@@ -137,7 +142,7 @@
   function handleActionsClose() {
     activeCreateAction.value = undefined;
     innerStep.value.createActionsVisible = false;
-    document.getElementById(innerStep.value.uniqueId.toString())?.click();
+    document.getElementById(`trigger-${innerStep.value.uniqueId}`)?.click();
   }
 </script>
