+++ frontend/src/views/project-management/projectAndPermission/member/index.vue
@@ -1,10 +1,10 @@
 <template>
   <div class="mb-4 grid grid-cols-4 gap-2">
-    <div class="col-span-2"
-      ><a-button v-permission="['PROJECT_USER:READ+ADD']" class="mr-3" type="primary" @click="addMember">{{
-        t('project.member.addMember')
-      }}</a-button></div
-    >
+    <div class="col-span-2">
+      <a-button v-permission="['PROJECT_USER:READ+ADD']" class="mr-3" type="primary" @click="addMember">
+        {{ t('project.member.addMember') }}
+      </a-button>
+    </div>
     <div>
       <a-select v-model="roleIds" @change="changeSelect">
         <a-option v-for="item of userGroupAll" :key="item.id" :value="item.id">{{ t(item.name) }}</a-option>
@@ -15,374 +15,45 @@
     >
     <div>
       <a-input-search
-        v-model="searchParams.keyword"
+        v-model="keyword"
         :max-length="255"
         :placeholder="t('project.member.searchMember')"
         allow-clear
         @search="searchHandler"
         @press-enter="searchHandler"
+        @clear="searchHandler"
       ></a-input-search
     ></div>
   </div>
-  <MsBaseTable
-    v-bind="propsRes"
+  <memberTable
+    v-if="isMounted"
     ref="memberTableRef"
-    :action-config="tableBatchActions"
-    @selected-change="handleTableSelect"
-    v-on="propsEvent"
-    @batch-action="handleTableBatch"
-  >
-    <template #userRole="{ record }">
-      <MsTagGroup
-        v-if="!record.showUserSelect"
-        :tag-list="record.userRoles || []"
-        type="primary"
-        theme="outline"
-        @click="changeUser(record)"
-      />
-      <a-select
-        v-else
-        v-model="record.selectUserList"
-        :popup-visible="record.showUserSelect"
-        multiple
-        class="w-[260px]"
-        :max-tag-count="2"
-        @popup-visible-change="(value) => userGroupChange(value, record)"
-      >
-        <a-option v-for="item of userGroupOptions" :key="item.id" :value="item.id">{{ item.name }}</a-option>
-      </a-select>
-    </template>
-    <template #enable="{ record }">
-      <div v-if="record.enable" class="flex items-center">
-        <icon-check-circle-fill class="mr-[2px] text-[rgb(var(--success-6))]" />
-        {{ t('organization.member.statusEnable') }}
-      </div>
-      <div v-else class="flex items-center text-[var(--color-text-4)]">
-        <MsIcon type="icon-icon_disable" class="mr-[2px]" />
-        {{ t('organization.member.statusDisable') }}
-      </div>
-    </template>
-    <template #operation="{ record }">
-      <MsRemoveButton
-        v-permission="['PROJECT_USER:READ+DELETE']"
-        position="br"
-        :title="t('project.member.deleteMemberTip', { name: characterLimit(record.name) })"
-        :sub-title-tip="t('project.member.subTitle')"
-        :loading="deleteLoading"
-        @ok="removeMember(record)"
-      />
-    </template>
-  </MsBaseTable>
-  <AddMemberModal
-    ref="projectMemberRef"
-    v-model:visible="addMemberVisible"
+    :keyword="keyword"
+    :role-ids="roleIds"
     :user-group-options="userGroupOptions"
-    @success="initData()"
-  />
-  <MsBatchModal
-    ref="batchModalRef"
-    v-model:visible="batchVisible"
-    :action="batchAction"
-    :select-data="selectData"
-    @add-user-group="addUserGroup"
   />
 </template>
 
 <script setup lang="ts">
+  import { getProjectUserGroup } from '@/api/modules/project-management/projectMember';
+  import { useI18n } from '@/hooks/useI18n';
+  import useAppStore from '@/store/modules/app';
+
+  import { ProjectUserOption } from '@/models/projectManagement/projectAndPermission';
+
+  const memberTable = defineAsyncComponent(() => import('./components/memberTable.vue'));
+
   /**
    * @description 项目管理-项目与权限-成员
    */
-  import { ref } from 'vue';
-  import { Message } from '@arco-design/web-vue';
-  import { isEqual } from 'lodash-es';
-
-  import MsBaseTable from '@/components/pure/ms-table/base-table.vue';
-  import type { BatchActionParams, BatchActionQueryParams, MsTableColumn } from '@/components/pure/ms-table/type';
-  import useTable from '@/components/pure/ms-table/useTable';
-  import MsTagGroup from '@/components/pure/ms-tag/ms-tag-group.vue';
-  import MsBatchModal from '@/components/business/ms-batch-modal/index.vue';
-  import MsRemoveButton from '@/components/business/ms-remove-button/MsRemoveButton.vue';
-  import AddMemberModal from './components/addMemberModal.vue';
-
-  import {
-    addOrUpdateProjectMember,
-    addProjectUserGroup,
-    batchRemoveMember,
-    getProjectMemberList,
-    getProjectUserGroup,
-    removeProjectMember,
-  } from '@/api/modules/project-management/projectMember';
-  import { useI18n } from '@/hooks/useI18n';
-  import useModal from '@/hooks/useModal';
-  import { useAppStore, useTableStore } from '@/store';
-  import { characterLimit, formatPhoneNumber } from '@/utils';
-  import { hasAnyPermission } from '@/utils/permission';
-
-  import type {
-    ActionProjectMember,
-    ProjectMemberItem,
-    ProjectUserOption,
-    SearchParams,
-  } from '@/models/projectManagement/projectAndPermission';
-  import { TableKeyEnum } from '@/enums/tableEnum';
-
-  const { t } = useI18n();
-  const { openModal } = useModal();
   const appStore = useAppStore();
-  const tableStore = useTableStore();
-  const lastProjectId = computed(() => appStore.currentProjectId);
-  const hasOperationPermission = computed(() => hasAnyPermission(['PROJECT_USER:READ+DELETE']));
-  const columns: MsTableColumn = [
-    {
-      title: 'project.member.tableColumnName',
-      slotName: 'name',
-      dataIndex: 'name',
-      showTooltip: true,
-      showDrag: false,
-    },
-    {
-      title: 'project.member.tableColumnEmail',
-      slotName: 'email',
-      dataIndex: 'email',
-      showTooltip: true,
-      showDrag: true,
-    },
-    {
-      title: 'project.member.tableColumnPhone',
-      slotName: 'phone',
-      dataIndex: 'phone',
-      showDrag: true,
-      width: 150,
-    },
-    {
-      title: 'project.member.tableColumnUserGroup',
-      slotName: 'userRole',
-      dataIndex: 'userRoleIdNameMap',
-      showDrag: true,
-      width: 300,
-    },
-    {
-      title: 'project.member.tableColumnStatus',
-      slotName: 'enable',
-      dataIndex: 'enable',
-      width: 150,
-    },
-    {
-      title: hasOperationPermission.value ? 'project.member.tableColumnActions' : '',
-      slotName: 'operation',
-      fixed: 'right',
-      dataIndex: 'operation',
-      width: hasOperationPermission.value ? 120 : 50,
-      showDrag: false,
-    },
-  ];
-
-  const tableBatchActions = {
-    baseAction: [
-      {
-        label: 'project.member.batchActionAddUserGroup',
-        eventTag: 'batchAddUserGroup',
-        permission: ['PROJECT_USER:READ+UPDATE'],
-      },
-      {
-        label: 'project.member.batchActionRemove',
-        eventTag: 'batchActionRemove',
-        permission: ['PROJECT_USER:READ+DELETE'],
-      },
-    ],
-  };
-
-  const tableSelected = ref<(string | number)[]>([]);
-
-  const handleTableSelect = (selectArr: (string | number)[]) => {
-    tableSelected.value = selectArr;
-  };
-
-  const { propsRes, propsEvent, loadList, setLoadListParams, resetSelector } = useTable(
-    getProjectMemberList,
-    {
-      tableKey: TableKeyEnum.PROJECT_MEMBER,
-      selectable: hasAnyPermission(['ORGANIZATION_MEMBER:READ+UPDATE']),
-      showSetting: true,
-      heightUsed: 288,
-      scroll: {
-        x: 1200,
-      },
-    },
-    (record) => {
-      return {
-        ...record,
-        phone: formatPhoneNumber(record.phone || ''),
-      };
-    }
-  );
-
-  const searchParams = ref<SearchParams>({
-    filter: {
-      roleIds: [],
-    },
-    projectId: lastProjectId.value,
-    keyword: '',
-  });
-
-  const roleIds = ref<string>('');
-  const initData = async () => {
-    setLoadListParams({ ...searchParams.value });
-    await loadList();
-  };
-
-  const searchHandler = () => {
-    initData();
-  };
-
-  const changeSelect = () => {
-    searchParams.value.filter.roleIds = roleIds.value ? [roleIds.value] : [];
-    initData();
-  };
-
-  // 跨页选择项
-  const selectData = ref<string[] | undefined>([]);
-
-  // 批量移除项目成员
-  const batchRemoveHandler = () => {
-    openModal({
-      type: 'error',
-      title: t('project.member.batchRemoveTip', { number: (selectData.value || []).length }),
-      content: t('project.member.batchRemoveContent'),
-      okText: t('project.member.deleteMemberConfirm'),
-      cancelText: t('project.member.Cancel'),
-      okButtonProps: {
-        status: 'danger',
-      },
-      onBeforeOk: async () => {
-        try {
-          const params: ActionProjectMember = {
-            projectId: lastProjectId.value,
-            userIds: selectData.value,
-          };
-          await batchRemoveMember(params);
-          Message.success(t('project.member.deleteMemberSuccess'));
-          loadList();
-          resetSelector();
-        } catch (error) {
-          console.log(error);
-        }
-      },
-      hideCancel: false,
-    });
-  };
-
-  // 移除项目成员
-  const deleteLoading = ref<boolean>(false);
-
-  const removeMember = async (record: ProjectMemberItem) => {
-    deleteLoading.value = true;
-    try {
-      if (lastProjectId.value && record.id) {
-        await removeProjectMember(lastProjectId.value, record.id);
-        Message.success(t('project.member.deleteMemberSuccess'));
-        initData();
-        resetSelector();
-      }
-    } catch (error) {
-      console.log(error);
-    } finally {
-      deleteLoading.value = false;
-    }
-  };
-
-  const batchVisible = ref<boolean>(false);
-  const batchAction = ref('');
-  const userGroupOptions = ref<ProjectUserOption[]>([]);
-  const batchModalRef = ref();
-
-  // 添加到用户组
-  const addUserGroup = async (target: string[]) => {
-    const params = {
-      projectId: lastProjectId.value,
-      userIds: selectData.value,
-      roleIds: target,
-    };
-    try {
-      await batchModalRef.value.batchRequestFun(addProjectUserGroup, params);
-      initData();
-      resetSelector();
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // 表格批量处理
-  const handleTableBatch = (event: BatchActionParams, params: BatchActionQueryParams) => {
-    selectData.value = params.selectedIds;
-    if (event.eventTag === 'batchActionRemove') {
-      batchRemoveHandler();
-    }
-    if (event.eventTag === 'batchAddUserGroup') {
-      batchVisible.value = true;
-      batchAction.value = event.eventTag;
-      batchModalRef.value.getTreeList(getProjectUserGroup, lastProjectId.value);
-    }
-  };
-
-  // 添加项目成员
-  const addMemberVisible = ref<boolean>(false);
-  const projectMemberRef = ref();
-
-  const addMember = () => {
-    addMemberVisible.value = true;
-    projectMemberRef.value.initProjectMemberOptions();
-  };
-
-  // 编辑项目成员
-  const editProjectMember = async (record: ProjectMemberItem) => {
-    const params: ActionProjectMember = {
-      projectId: lastProjectId.value,
-      userId: record.id,
-      roleIds: record.selectUserList,
-    };
-    try {
-      await addOrUpdateProjectMember(params);
-      Message.success(t('project.member.batchUpdateSuccess'));
-      record.showUserSelect = false;
-      loadList();
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // 项目用户组改变回调
-  const changeUser = (record: ProjectMemberItem) => {
-    if (!hasAnyPermission(['PROJECT_USER:READ+UPDATE'])) {
-      return;
-    }
-    if (record.enable) {
-      record.showUserSelect = true;
-      record.selectUserList = (record.userRoles || []).map((item) => item.id);
-    }
-  };
-
-  // 面板改变的回调
-  const userGroupChange = (visible: boolean, record: ProjectMemberItem) => {
-    if (visible) {
-      return;
-    }
-    if ((record.selectUserList || []).length < 1) {
-      Message.warning(t('project.member.selectUserEmptyTip'));
-      return;
-    }
-    const userGroupIds = (record.userRoles || []).map((item: any) => item.id);
-    if (isEqual(userGroupIds, record.selectUserList)) {
-      record.showUserSelect = false;
-      return;
-    }
-    editProjectMember(record);
-  };
+  const { t } = useI18n();
 
   const userGroupAll = ref<ProjectUserOption[]>([]);
+  const userGroupOptions = ref<ProjectUserOption[]>([]);
 
   const initOptions = async () => {
-    userGroupOptions.value = await getProjectUserGroup(lastProjectId.value);
+    userGroupOptions.value = await getProjectUserGroup(appStore.currentProjectId);
     userGroupAll.value = [
       {
         id: '',
@@ -391,14 +62,32 @@
       ...userGroupOptions.value,
     ];
   };
-  const memberTableRef = ref();
   initOptions();
+  const memberTableRef = ref<InstanceType<typeof memberTable> | null>(null);
 
-  onBeforeMount(() => {
+  const roleIds = ref<string>('');
+  const keyword = ref<string>('');
+  const initData = async () => {
+    memberTableRef.value?.initData();
+  };
+
+  const searchHandler = () => {
     initData();
-  });
+  };
 
-  await tableStore.initColumn(TableKeyEnum.PROJECT_MEMBER, columns, 'drawer');
+  const changeSelect = () => {
+    initData();
+  };
+
+  function addMember() {
+    memberTableRef.value?.addMember();
+  }
+
+  const isMounted = ref(false);
+
+  onMounted(() => {
+    isMounted.value = true; // 解决memberTable因为异步初始化列导致的路由组件渲染两次的问题
+  });
 </script>
 
 <style scoped></style>
